Got channel abstractions working
Added in and out channels wrapping
queues.  Impemented dmap!, a
channel based version of dmap
with a work-stealing queue setup.

Need to unify the representation of
object id's, since we are getting
different results for keywords
and strings and it's messing us
up.

Channel closure in the cluster map
is still not being tracked correctly,
but the channel is closed.