modified to use delays
We now default to delayed initialization
and provide a get-cluster and get-client
api that is aware of delayed cluster
instances.

We use this (in the implementation)
instead of directly referencing
*cluster* or *client* to
enable the default autoconfig
to work, but without impacting
aot compilation (hopefully).

callers can still define their
own cluster instance via dynvars
and it works out of the box.